docs: basic comments added
